changed the arguments of encode to double

+++ qam.h
@@ -25,7 +25,7 @@ namespace halow
             QAM(int numBits, double power, int gain =0 );
 
             //member function definition
-            void encode(const char *data, int *cPoint ); //calculate constellation point from the bit vector
+            void encode(const char *data, double *cPoint ); //calculate constellation point from the bit vector
             void decode(const char* IQ); //retrieves bits from the received constellation points
     };
 }
